Round where the pin rounds, at every spelling
/simplify over the wave, applied. Three of the findings were the same
defect the wave itself fixed, half-applied somewhere else.

The build breaker first: `transformed_vertices` accumulated a float lane
with the record's now-double translation, which is C4244 under /W4 /WX.
Every scene would have failed to build with MSVC, not only the
floating-origin ones; clang-cl accepts it silently, which is why it
landed. It also double-rounded the ordinary bake, where the pin's
`allocateMat4()` is a Float32Array and its own GPU multiply sees the
narrowed translation -- so the cast restores the pin's rounding point
rather than merely silencing a warning.

Then the two half-applications: `mesh.position.x = v` still narrowed and
widened the rounded value straight back into the double field, on a path
five registered scenes reach; and the physics pose kept its three
`static_cast<float>` inside the widened literal, which the new assertion
green-lit by matching only the type name.

And a correction to this branch's own earlier change: the positional
light offset had been moved to `light.local_matrix` on the claim that it
matched the pin and the writer beside it. Neither holds -- the entry
writer composes its own local matrix from `light.position`, and the glTF
punctual-light emission never writes `local_matrix` at all, so an
imported light under floating origin would have been placed at the
origin. It reads `light.position` again.

`renderer:high-precision-matrix` is a reached feature again. It was
removed as gating nothing, and it now gates the width every composed
camera matrix is stored at -- `createEngine` accepts the flag on its own,
and a scene asking for it was getting a float world where the pin's
allocator has swapped to F64.

Two width limits are filed rather than applied, both blocked on
measurements no reached scene provides: a billboard's anchor is still
stored in a float vector, and the matrices the pin's process-global
allocator widens beyond the camera's world and a node's translation.

+++ src/lowering/physics-lowerer.ts
@@ -759,9 +759,9 @@ void sync_body_to_node(Engine& engine, const PhysicsBody& body) {
         pal::physics_body_get_transform(body.handle);
     MeshRecord& mesh = engine.meshes[body.node.value];
     mesh.position = Vec3d{
-        static_cast<float>(transform.position[0]),
-        static_cast<float>(transform.position[1]),
-        static_cast<float>(transform.position[2]),
+        transform.position[0],
+        transform.position[1],
+        transform.position[2],
     };
     mesh.rotation_quaternion = Vec4{
         static_cast<float>(transform.rotation[0]),
